% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_functions}
\alias{glm_functions}
\title{GLM Functions for Distance-Dependent Expression Analysis}
\description{
Core statistical functions for the RIPPLE pipeline. Fits Poisson
generalized linear models to test whether gene expression changes as a
function of distance from query cells.
}
