% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_adapter.R
\name{input_adapter}
\alias{input_adapter}
\title{Input Adapter}
\description{
Internal function for normalizing input data from multiple
canonical classes (Seurat, SingleCellExperiment, SpatialExperiment, or
an \code{.rds} file containing one of these) into a common structure used by
the RIPPLE pipeline.
}
\keyword{internal}
