% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{make_input}
\alias{make_input}
\title{Input Constructors}
\description{
User-facing constructors for building a RIPPLE-compatible
canonical object (Seurat or SpatialExperiment) from raw matrices or CSVs.
}
