% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation}
\alias{permutation}
\title{Permutation Testing Functions}
\description{
Functions for label-permutation testing to validate query cell
specificity of distance-expression gradients, and for merging GPU
permutation results.
}
